fix possible cross-platform issue that code failing to pass on some tests

+++ src/uucore/src/lib/features/format/spec.rs
@@ -374,7 +374,7 @@ impl Spec {
                 let precision = resolve_asterisk(*precision, &mut args)?.unwrap_or(0);
                 let i = args.get_i64();
 
-                if precision > i32::MAX as usize {
+                if precision as u64 > i32::MAX as u64 {
                     return Err(FormatError::InvalidPrecision(precision.to_string()));
                 }
 
@@ -397,7 +397,7 @@ impl Spec {
                 let precision = resolve_asterisk(*precision, &mut args)?.unwrap_or(0);
                 let i = args.get_u64();
 
-                if precision > i32::MAX as usize {
+                if precision as u64 > i32::MAX as u64 {
                     return Err(FormatError::InvalidPrecision(precision.to_string()));
                 }
 
@@ -423,7 +423,7 @@ impl Spec {
                 let precision = resolve_asterisk(*precision, &mut args)?.unwrap_or(6);
                 let f = args.get_f64();
 
-                if precision > i32::MAX as usize {
+                if precision as u64 > i32::MAX as u64 {
                     return Err(FormatError::InvalidPrecision(precision.to_string()));
                 }
